Changed version comparison method
Sometimes it is not required to update both: android app and arduino
code, but because the version comparison just checked if versions of
both were the same, the version names of both had to be changed
(otherwise the notification about different versions would be
displayed).

For that reason I changed it so the user will be notified only when it
is really required to also update the Arduino code (or only the mobile
app).

From now: 1.08 is compatible with 1.081 or 1.83148, but it's not compatible with 1.09 or 1.091319 (in such case notification will be displayed in the app).

Let's say that the version of Arduino code and mobile app is 1.08 and
I'd like to make only small change to the app that doesn't require
updating Arduino code. I'd set the version of mobile app to be 1.081,
despite not being identical to the Arduino code version, the
notification about different versions will not be displayed.

Until now it possibly was pain in the ass to additionally update Arduino
code. Hopefully now it will become more convenient to keep the project
up to date.